Icons updated to JSX
