+++ src/assets/allIcons.jsx
@@ -1,5 +1,23 @@
 /* eslint-disable react/prop-types */
 
+export const ArrowRightIcon = ({ className, ...rest }) => (
+  <svg 
+    xmlns="http://www.w3.org/2000/svg" 
+    fill="currentColor" 
+    viewBox="0 0 24 24" 
+    strokeWidth={1.5} 
+    stroke="currentColor" 
+    {...rest}
+    className={`w-full h-auto ${className}`}
+  >
+    <path 
+      fillRule="evenodd"
+      d="M12 2.25c-5.385 0-9.75 4.365-9.75 9.75s4.365 9.75 9.75 9.75 9.75-4.365 9.75-9.75S17.385 2.25 12 2.25zm4.28 10.28a.75.75 0 000-1.06l-3-3a.75.75 0 10-1.06 1.06l1.72 1.72H8.25a.75.75 0 000 1.5h5.69l-1.72 1.72a.75.75 0 101.06 1.06l3-3z" 
+      clipRule="evenodd" 
+    />
+  </svg>
+);
+
 export const ChevronLeftIcon = ({ className, ...rest }) => (
   <svg 
     xmlns="http://www.w3.org/2000/svg" 
@@ -12,7 +30,6 @@ export const ChevronLeftIcon = ({ className, ...rest }) => (
   >
     <path strokeLinecap="round" strokeLinejoin="round" d="M18.75 19.5l-7.5-7.5 7.5-7.5m-6 15L5.25 12l7.5-7.5" />
   </svg>
-
 );
 
 export const ChevronRightIcon = ({ className, ...rest }) => (
@@ -27,65 +44,101 @@ export const ChevronRightIcon = ({ className, ...rest }) => (
   >
     <path strokeLinecap="round" strokeLinejoin="round" d="M11.25 4.5l7.5 7.5-7.5 7.5m-6-15l7.5 7.5-7.5 7.5" />
   </svg>
-
-
 );
 
-export const HomeIcon = ({ className, ...rest }) => (
+export const TruckIcon = ({ className, ...rest }) => (
   <svg 
     xmlns="http://www.w3.org/2000/svg" 
-    fill="none" 
+    fill="currentColor" 
     viewBox="0 0 24 24" 
-    strokeWidth={1.5} 
-    stroke="currentColor" 
     {...rest}
     className={`w-full h-auto ${className}`}
   >
-    <path strokeLinecap="round" strokeLinejoin="round" d="M2.25 12l8.954-8.955c.44-.439 1.152-.439 1.591 0L21.75 12M4.5 9.75v10.125c0 .621.504 1.125 1.125 1.125H9.75v-4.875c0-.621.504-1.125 1.125-1.125h2.25c.621 0 1.125.504 1.125 1.125V21h4.125c.621 0 1.125-.504 1.125-1.125V9.75M8.25 21h8.25" />
+    <path d="M3.375 4.5C2.339 4.5 1.5 5.34 1.5 6.375V13.5h12V6.375c0-1.036-.84-1.875-1.875-1.875h-8.25zM13.5 15h-12v2.625c0 1.035.84 1.875 1.875 1.875h.375a3 3 0 116 0h3a.75.75 0 00.75-.75V15z" />
+    <path d="M8.25 19.5a1.5 1.5 0 10-3 0 1.5 1.5 0 003 0zM15.75 6.75a.75.75 0 00-.75.75v11.25c0 .087.015.17.042.248a3 3 0 015.958.464c.853-.175 1.522-.935 1.464-1.883a18.659 18.659 0 00-3.732-10.104 1.837 1.837 0 00-1.47-.725H15.75z" />
+    <path d="M19.5 19.5a1.5 1.5 0 10-3 0 1.5 1.5 0 003 0z" />
   </svg>
 );
 
-export const ContactIcon = ({ className, ...rest }) => (
+export const ShieldIcon = ({ className, ...rest }) => (
   <svg 
     xmlns="http://www.w3.org/2000/svg" 
-    fill="none" 
+    fill="currentColor" 
     viewBox="0 0 24 24" 
-    strokeWidth={1.5} 
-    stroke="currentColor" 
     {...rest}
     className={`w-full h-auto ${className}`}
   >
-    <path strokeLinecap="round" strokeLinejoin="round" d="M21.75 6.75v10.5a2.25 2.25 0 01-2.25 2.25h-15a2.25 2.25 0 01-2.25-2.25V6.75m19.5 0A2.25 2.25 0 0019.5 4.5h-15a2.25 2.25 0 00-2.25 2.25m19.5 0v.243a2.25 2.25 0 01-1.07 1.916l-7.5 4.615a2.25 2.25 0 01-2.36 0L3.32 8.91a2.25 2.25 0 01-1.07-1.916V6.75" />
+    <path 
+      fillRule="evenodd" 
+      d="M12.516 2.17a.75.75 0 00-1.032 0 11.209 11.209 0 01-7.877 3.08.75.75 0 00-.722.515A12.74 12.74 0 002.25 9.75c0 5.942 4.064 10.933 9.563 12.348a.749.749 0 00.374 0c5.499-1.415 9.563-6.406 9.563-12.348 0-1.39-.223-2.73-.635-3.985a.75.75 0 00-.722-.516l-.143.001c-2.996 0-5.717-1.17-7.734-3.08zm3.094 8.016a.75.75 0 10-1.22-.872l-3.236 4.53L9.53 12.22a.75.75 0 00-1.06 1.06l2.25 2.25a.75.75 0 001.14-.094l3.75-5.25z" 
+      clipRule="evenodd" 
+    />
   </svg>
 );
 
-export const PhotosIcon = ({ className, ...rest }) => (
+export const LoveIcon = ({ className, ...rest }) => (
   <svg 
     xmlns="http://www.w3.org/2000/svg" 
-    fill="none" 
+    fill="currentColor" 
     viewBox="0 0 24 24" 
-    strokeWidth={1.5} 
-    stroke="currentColor" 
     {...rest}
     className={`w-full h-auto ${className}`}
   >
-    <path strokeLinecap="round" strokeLinejoin="round" d="M6.827 6.175A2.31 2.31 0 015.186 7.23c-.38.054-.757.112-1.134.175C2.999 7.58 2.25 8.507 2.25 9.574V18a2.25 2.25 0 002.25 2.25h15A2.25 2.25 0 0021.75 18V9.574c0-1.067-.75-1.994-1.802-2.169a47.865 47.865 0 00-1.134-.175 2.31 2.31 0 01-1.64-1.055l-.822-1.316a2.192 2.192 0 00-1.736-1.039 48.774 48.774 0 00-5.232 0 2.192 2.192 0 00-1.736 1.039l-.821 1.316z" />
-    <path strokeLinecap="round" strokeLinejoin="round" d="M16.5 12.75a4.5 4.5 0 11-9 0 4.5 4.5 0 019 0zM18.75 10.5h.008v.008h-.008V10.5z" />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M9.23576 0.0195141C8.27717 0.167622 7.66306 0.45699 7.06498 1.04233C5.56279 2.51266 5.47498 4.91504 6.85519 6.78199C6.9655 6.93126 7.31946 7.32504 7.64177 7.65709C8.66521 8.71149 10.0762 9.8829 12.1334 11.3862L12.7104 11.8078L13.3758 11.3146C15.4311 9.79122 16.7986 8.65449 17.7468 7.6812C18.4867 6.92175 18.779 6.53034 19.0798 5.89619C19.4122 5.19519 19.5241 4.69502 19.522 3.91884C19.52 3.15071 19.443 2.80603 19.1195 2.11478C18.9137 1.67508 18.8313 1.55779 18.4554 1.16941C17.9748 0.672955 17.5594 0.396078 17.0182 0.21151C16.1133 -0.0971494 15.2038 -0.0512363 14.3696 0.345246C13.8865 0.574859 13.6371 0.760103 13.1394 1.25878L12.7064 1.69259L12.2252 1.21113C11.6926 0.678356 11.2714 0.396319 10.7313 0.210931C10.2294 0.0387088 9.61809 -0.0395652 9.23576 0.0195141ZM7.7121 13.547C7.11299 13.6002 6.25303 13.7491 5.65154 13.904C4.95984 14.082 4.03186 14.4079 4.03186 14.4727C4.03186 14.5536 6.38084 21.8036 6.42228 21.8506C6.44131 21.8722 6.676 21.8521 6.94384 21.8059C7.37516 21.7316 7.97468 21.7186 12.1893 21.6921L16.9479 21.6622L17.3988 21.5362C17.946 21.3833 18.4694 21.1403 18.8666 20.8547C19.0297 20.7374 20.1602 19.7775 21.379 18.7216C22.9433 17.3663 23.6347 16.7371 23.7306 16.5817C24.1026 15.9784 24.0886 15.3468 23.6933 14.8945C23.5794 14.7642 23.4001 14.623 23.289 14.5763C23.0429 14.4726 22.6293 14.4667 22.4075 14.5637C22.316 14.6037 21.5671 15.0658 20.7432 15.5905C19.0719 16.6548 18.5405 16.9463 17.8454 17.1799C16.7076 17.5624 16.3503 17.5896 12.4589 17.5906L9.18888 17.5915V16.8701V16.1487L11.5916 16.1346C13.9158 16.121 14.0012 16.1172 14.2053 16.0196C14.5171 15.8705 14.701 15.6897 14.8416 15.3941C15.0971 14.8568 14.9636 14.1955 14.5209 13.8057C14.1624 13.4901 14.1773 13.4914 11.0205 13.4987C9.45873 13.5024 7.96995 13.5241 7.7121 13.547ZM1.32545 15.6528L0.00210969 16.2652L0.00103141 20.1326L0 24L0.503982 23.8168C0.781196 23.7161 1.89403 23.3232 2.97701 22.9438C4.05999 22.5643 4.96425 22.2376 4.98647 22.2177C5.02923 22.1795 2.73369 15.0336 2.68011 15.0379C2.6629 15.0393 2.05334 15.316 1.32545 15.6528Z"
+    />
   </svg>
 );
 
-export const SettingsIcon = ({ className, ...rest }) => (
-  <svg 
-    xmlns="http://www.w3.org/2000/svg" 
-    fill="none" 
-    viewBox="0 0 24 24" 
-    strokeWidth={1.5} 
-    stroke="currentColor" 
+export const TwitterIcon = ({ className, ...rest }) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={24}
+    height={24}
+    viewBox="0 0 24 24"
+    fill="currentColor"
     {...rest}
     className={`w-full h-auto ${className}`}
   >
-    <path strokeLinecap="round" strokeLinejoin="round" d="M9.594 3.94c.09-.542.56-.94 1.11-.94h2.593c.55 0 1.02.398 1.11.94l.213 1.281c.063.374.313.686.645.87.074.04.147.083.22.127.324.196.72.257 1.075.124l1.217-.456a1.125 1.125 0 011.37.49l1.296 2.247a1.125 1.125 0 01-.26 1.431l-1.003.827c-.293.24-.438.613-.431.992a6.759 6.759 0 010 .255c-.007.378.138.75.43.99l1.005.828c.424.35.534.954.26 1.43l-1.298 2.247a1.125 1.125 0 01-1.369.491l-1.217-.456c-.355-.133-.75-.072-1.076.124a6.57 6.57 0 01-.22.128c-.331.183-.581.495-.644.869l-.213 1.28c-.09.543-.56.941-1.11.941h-2.594c-.55 0-1.02-.398-1.11-.94l-.213-1.281c-.062-.374-.312-.686-.644-.87a6.52 6.52 0 01-.22-.127c-.325-.196-.72-.257-1.076-.124l-1.217.456a1.125 1.125 0 01-1.369-.49l-1.297-2.247a1.125 1.125 0 01.26-1.431l1.004-.827c.292-.24.437-.613.43-.992a6.932 6.932 0 010-.255c.007-.378-.138-.75-.43-.99l-1.004-.828a1.125 1.125 0 01-.26-1.43l1.297-2.247a1.125 1.125 0 011.37-.491l1.216.456c.356.133.751.072 1.076-.124.072-.044.146-.087.22-.128.332-.183.582-.495.644-.869l.214-1.281z" />
-    <path strokeLinecap="round" strokeLinejoin="round" d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />
+    <path
+      d="M22.5 5.776a8.902 8.902 0 0 1-2.474.65 4.18 4.18 0 0 0 1.894-2.28 8.922 8.922 0 0 1-2.737 1 4.387 4.387 0 0 0-3.145-1.302c-2.379 0-4.308 1.847-4.308 4.124 0 .323.038.638.112.94-3.58-.173-6.754-1.815-8.879-4.31A3.961 3.961 0 0 0 2.38 6.67c0 1.431.762 2.694 1.916 3.433a4.455 4.455 0 0 1-1.95-.518v.051c0 1.998 1.485 3.666 3.454 4.045-.36.093-.741.145-1.135.145a4.31 4.31 0 0 1-.81-.077c.548 1.64 2.138 2.832 4.023 2.865a8.89 8.89 0 0 1-5.35 1.763c-.348 0-.69-.02-1.028-.057a12.572 12.572 0 0 0 6.603 1.856c7.925 0 12.257-6.285 12.257-11.736l-.015-.534a8.412 8.412 0 0 0 2.155-2.13Z"
+    />
+  </svg>
+);
+
+export const FacebookIcon = ({ className, ...rest }) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={24}
+    height={24}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    {...rest}
+    className={`w-full h-auto ${className}`}
+  >
+    <path
+      d="M17.565 1.005 14.787 1c-3.12 0-5.136 2.125-5.136 5.414v2.497H6.859c-.242 0-.437.2-.437.448v3.617c0 .248.196.449.437.449H9.65v9.126c0 .248.195.449.437.449h3.643a.443.443 0 0 0 .437-.449v-9.126h3.265c.241 0 .436-.2.436-.449l.002-3.617a.455.455 0 0 0-.128-.317.431.431 0 0 0-.31-.131h-3.265V6.794c0-1.017.236-1.533 1.526-1.533h1.87a.444.444 0 0 0 .437-.45V1.454a.443.443 0 0 0-.436-.448Z"
+    />
+  </svg>
+);
+
+export const InstagramIcon = ({ className, ...rest }) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={24}
+    height={24}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    {...rest}
+    className={`w-full h-auto ${className}`}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M2 12c0-4.191 0-6.287 1.05-7.733A5.5 5.5 0 0 1 4.267 3.05C5.713 2 7.81 2 12 2s6.287 0 7.733 1.05a5.5 5.5 0 0 1 1.217 1.217C22 5.713 22 7.81 22 12s0 6.287-1.05 7.733a5.5 5.5 0 0 1-1.217 1.217C18.287 22 16.19 22 12 22s-6.287 0-7.733-1.05a5.5 5.5 0 0 1-1.217-1.217C2 18.287 2 16.19 2 12Zm10.001-5.297a5.294 5.294 0 1 0 0 10.589 5.294 5.294 0 0 0 0-10.589Zm0 8.603a3.313 3.313 0 0 1-3.309-3.309 3.313 3.313 0 0 1 3.31-3.309 3.313 3.313 0 0 1 3.308 3.31 3.313 3.313 0 0 1-3.309 3.308Zm6.394-8.999a.705.705 0 1 1-1.41 0 .705.705 0 0 1 1.41 0Z"
+    />
   </svg>
 );
 
@@ -106,112 +159,19 @@ export const GithubIcon = ({ className, ...rest }) => (
   </svg>
 );
 
-export const TwitterIcon = ({ className, ...rest }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="1em"
-    height="1em"
-    viewBox="0 0 256 209"
+export const CopyrightIcon = ({ className, ...rest }) => (
+  <svg 
+    xmlns="http://www.w3.org/2000/svg" 
+    fill="currentColor" 
+    viewBox="0 0 24 24" 
     {...rest}
     className={`w-full h-auto ${className}`}
   >
-    <path fill="none" d="M0 0h256v209H0z" />
     <path
-      fill="#55acee"
-      d="M256 25.45a105.04 105.04 0 0 1-30.166 8.27c10.845-6.5 19.172-16.793 23.093-29.057a105.183 105.183 0 0 1-33.351 12.745C205.995 7.201 192.346.822 177.239.822c-29.006 0-52.523 23.516-52.523 52.52 0 4.117.465 8.125 1.36 11.97-43.65-2.191-82.35-23.1-108.255-54.876-4.52 7.757-7.11 16.78-7.11 26.404 0 18.222 9.273 34.297 23.365 43.716a52.312 52.312 0 0 1-23.79-6.57c-.003.22-.003.44-.003.661 0 25.447 18.104 46.675 42.13 51.5a52.592 52.592 0 0 1-23.718.9c6.683 20.866 26.08 36.05 49.062 36.475-17.975 14.086-40.622 22.483-65.228 22.483-4.24 0-8.42-.249-12.529-.734 23.243 14.902 50.85 23.597 80.51 23.597 96.607 0 149.434-80.031 149.434-149.435 0-2.278-.05-4.543-.152-6.795A106.748 106.748 0 0 0 256 25.45"
+      d="M10.688 16.096c-1.11 0-2.112-.245-3.008-.736a5.48 5.48 0 0 1-2.096-2.032c-.501-.875-.752-1.85-.752-2.928 0-1.077.25-2.048.752-2.912A5.415 5.415 0 0 1 7.696 5.44c.896-.49 1.899-.736 3.008-.736.832 0 1.6.139 2.304.416A4.776 4.776 0 0 1 14.8 6.336l-.752.752c-.875-.885-1.979-1.328-3.312-1.328-.885 0-1.69.203-2.416.608a4.501 4.501 0 0 0-1.712 1.664C6.203 8.736 6 9.525 6 10.4c0 .875.203 1.664.608 2.368a4.501 4.501 0 0 0 1.712 1.664c.725.405 1.53.608 2.416.608 1.344 0 2.448-.448 3.312-1.344l.752.752a4.912 4.912 0 0 1-1.808 1.232 6.232 6.232 0 0 1-2.304.416Z"
+      fill="currentColor"
     />
-  </svg>
-);
-
-export const InstagramIcon = ({ className, ...rest }) => (
-  <svg
-    width="1em"
-    height="1em"
-    viewBox="0 0 32 32"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...rest}
-    className={`w-full h-auto ${className}`}
-  >
-    <rect
-      x={2}
-      y={2}
-      width={28}
-      height={28}
-      rx={6}
-      fill="url(#paint0_radial_87_7153)"
-    />
-    <rect
-      x={2}
-      y={2}
-      width={28}
-      height={28}
-      rx={6}
-      fill="url(#paint1_radial_87_7153)"
-    />
-    <rect
-      x={2}
-      y={2}
-      width={28}
-      height={28}
-      rx={6}
-      fill="url(#paint2_radial_87_7153)"
-    />
-    <path
-      d="M23 10.5C23 11.3284 22.3284 12 21.5 12C20.6716 12 20 11.3284 20 10.5C20 9.67157 20.6716 9 21.5 9C22.3284 9 23 9.67157 23 10.5Z"
-      fill="white"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M16 21C18.7614 21 21 18.7614 21 16C21 13.2386 18.7614 11 16 11C13.2386 11 11 13.2386 11 16C11 18.7614 13.2386 21 16 21ZM16 19C17.6569 19 19 17.6569 19 16C19 14.3431 17.6569 13 16 13C14.3431 13 13 14.3431 13 16C13 17.6569 14.3431 19 16 19Z"
-      fill="white"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M6 15.6C6 12.2397 6 10.5595 6.65396 9.27606C7.2292 8.14708 8.14708 7.2292 9.27606 6.65396C10.5595 6 12.2397 6 15.6 6H16.4C19.7603 6 21.4405 6 22.7239 6.65396C23.8529 7.2292 24.7708 8.14708 25.346 9.27606C26 10.5595 26 12.2397 26 15.6V16.4C26 19.7603 26 21.4405 25.346 22.7239C24.7708 23.8529 23.8529 24.7708 22.7239 25.346C21.4405 26 19.7603 26 16.4 26H15.6C12.2397 26 10.5595 26 9.27606 25.346C8.14708 24.7708 7.2292 23.8529 6.65396 22.7239C6 21.4405 6 19.7603 6 16.4V15.6ZM15.6 8H16.4C18.1132 8 19.2777 8.00156 20.1779 8.0751C21.0548 8.14674 21.5032 8.27659 21.816 8.43597C22.5686 8.81947 23.1805 9.43139 23.564 10.184C23.7234 10.4968 23.8533 10.9452 23.9249 11.8221C23.9984 12.7223 24 13.8868 24 15.6V16.4C24 18.1132 23.9984 19.2777 23.9249 20.1779C23.8533 21.0548 23.7234 21.5032 23.564 21.816C23.1805 22.5686 22.5686 23.1805 21.816 23.564C21.5032 23.7234 21.0548 23.8533 20.1779 23.9249C19.2777 23.9984 18.1132 24 16.4 24H15.6C13.8868 24 12.7223 23.9984 11.8221 23.9249C10.9452 23.8533 10.4968 23.7234 10.184 23.564C9.43139 23.1805 8.81947 22.5686 8.43597 21.816C8.27659 21.5032 8.14674 21.0548 8.0751 20.1779C8.00156 19.2777 8 18.1132 8 16.4V15.6C8 13.8868 8.00156 12.7223 8.0751 11.8221C8.14674 10.9452 8.27659 10.4968 8.43597 10.184C8.81947 9.43139 9.43139 8.81947 10.184 8.43597C10.4968 8.27659 10.9452 8.14674 11.8221 8.0751C12.7223 8.00156 13.8868 8 15.6 8Z"
-      fill="white"
-    />
-    <defs>
-      <radialGradient
-        id="paint0_radial_87_7153"
-        cx={0}
-        cy={0}
-        r={1}
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(12 23) rotate(-55.3758) scale(25.5196)"
-      >
-        <stop stopColor="#B13589" />
-        <stop offset={0.79309} stopColor="#C62F94" />
-        <stop offset={1} stopColor="#8A3AC8" />
-      </radialGradient>
-      <radialGradient
-        id="paint1_radial_87_7153"
-        cx={0}
-        cy={0}
-        r={1}
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(11 31) rotate(-65.1363) scale(22.5942)"
-      >
-        <stop stopColor="#E0E8B7" />
-        <stop offset={0.444662} stopColor="#FB8A2E" />
-        <stop offset={0.71474} stopColor="#E2425C" />
-        <stop offset={1} stopColor="#E2425C" stopOpacity={0} />
-      </radialGradient>
-      <radialGradient
-        id="paint2_radial_87_7153"
-        cx={0}
-        cy={0}
-        r={1}
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(0.500002 3) rotate(-8.1301) scale(38.8909 8.31836)"
-      >
-        <stop offset={0.156701} stopColor="#406ADC" />
-        <stop offset={0.467799} stopColor="#6A45BE" />
-        <stop offset={1} stopColor="#6A45BE" stopOpacity={0} />
-      </radialGradient>
-    </defs>
+    <circle cx={10} cy={10} r={9.5} stroke="currentColor" />
   </svg>
 );
