--- src/assets/icons/index.js
@@ -1,27 +1,0 @@
-import arrowRight from './arrow-right.svg'
-import chevronLeft from './chevron-left.svg'
-import chevronRight from './chevron-right.svg'
-import copyrightSign from './copyright-sign.svg'
-import facebook from './facebook.svg'
-import hamburger from './hamburger.svg'
-import instagram from './instagram.svg'
-import support from './support.svg'
-import shieldTick from './shield-tick.svg'
-import star from './star.svg'
-import truckFast from './truck-fast.svg'
-import twitter from './twitter.svg'
-
-export {
-    arrowRight,
-    chevronRight,
-    chevronLeft,
-    copyrightSign,
-    facebook,
-    hamburger,
-    instagram,
-    support,
-    shieldTick,
-    star,
-    truckFast,
-    twitter
-}
