+++ src/components/Button.jsx
@@ -1,4 +1,5 @@
 const Button = ({
+  children,
   label,
   iconURL,
   backgroundColor,
@@ -8,7 +9,7 @@ const Button = ({
 }) => {
   return (
     <button
-      className={`flex justify-center items-center gap-2 px-7 py-4 border font-montserrat text-lg leading-none
+      className={`flex justify-center items-center gap-4 px-7 py-4 border font-montserrat text-lg leading-none
       ${
         backgroundColor
           ? `${backgroundColor} ${textColor} ${borderColor}`
@@ -17,6 +18,7 @@ const Button = ({
     >
       {label}
 
+      {/* Use the iconUrl if i want to use an svg icon */}
       {iconURL && (
         <img
           src={iconURL}
@@ -24,6 +26,9 @@ const Button = ({
           className='ml-2 rounded-full bg-white w-5 h-5'
         />
       )}
+      
+      {/* Use this instead of iconUrl if i want to use a jsx icon component */}
+      {children}
     </button>
   );
 };
