+++ src/constants/data.jsx
@@ -1,0 +1,152 @@
+import { bigShoe1, bigShoe2, bigShoe3, customer1, customer2, shoe4, shoe5, shoe6, shoe7, thumbnailShoe1, thumbnailShoe2, thumbnailShoe3 } from "../assets/images";
+import { TruckIcon, ShieldIcon, LoveIcon, TwitterIcon, FacebookIcon, InstagramIcon } from "../assets/allIcons";
+
+export const ourServices = [
+  {
+    imgURL: <TruckIcon className={"w-6 h-6 fill-white"} />,
+    label: "Free shipping",
+    subtext: "Enjoy seamless shopping with our complimentary shipping service.",
+  },
+  {
+    imgURL: <ShieldIcon className={"w-6 h-6 fill-white"} />,
+    label: "Secure Payment",
+    subtext: "Experience worry-free transactions with our secure payment options.",
+  },
+  {
+    imgURL: <LoveIcon className={"w-6 h-6 fill-white"} />,
+    label: "Love to help you",
+    subtext: "Our dedicated team is here to assist you every step of the way.",
+  },
+];
+
+export const navLinks = [
+  { href: "#home", label: "Home" },
+  { href: "#about-us", label: "About Us" },
+  { href: "#products", label: "Products" },
+  { href: "#contact-us", label: "Contact Us" },
+];
+
+export const shoes = [
+  {
+    thumbnail: thumbnailShoe1,
+    bigShoe: bigShoe1,
+  },
+  {
+    thumbnail: thumbnailShoe2,
+    bigShoe: bigShoe2,
+  },
+  {
+    thumbnail: thumbnailShoe3,
+    bigShoe: bigShoe3,
+  },
+];
+
+export const statistics = [
+  { value: '1k+', label: 'Brands' },
+  { value: '500+', label: 'Shops' },
+  { value: '250k+', label: 'Customers' },
+];
+
+export const products = [
+  {
+    imgURL: shoe4,
+    name: "Nike Air Jordan-01",
+    price: "$200.20",
+    rating: 4.5,
+  },
+  {
+    imgURL: shoe5,
+    name: "Nike Air Jordan-10",
+    price: "$210.20",
+    rating: 4,
+  },
+  {
+    imgURL: shoe6,
+    name: "Nike Air Jordan-100",
+    price: "$220.20",
+    rating: 5,
+  },
+  {
+    imgURL: shoe7,
+    name: "Nike Air Jordan-001",
+    price: "$230.20",
+    rating: 3.5,
+  },
+];
+
+export const reviews = [
+  {
+    imgURL: customer1,
+    customerName: 'Morich Brown',
+    rating: 4,
+    feedback: "1 The attention to detail and the quality of the product exceeded my expectations. Highly recommended!"
+  },
+  {
+    imgURL: customer2,
+    customerName: 'Lota Mongeskar',
+    rating: 4.5,
+    feedback: "2 The product not only met but exceeded my expectations. I'll definitely be a returning customer!"
+  },
+  {
+    imgURL: customer2,
+    customerName: 'Lota Mongeskar',
+    rating: 5,
+    feedback: "3 The product not only met but exceeded my expectations. I'll definitely be a returning customer!"
+  },
+  {
+    imgURL: customer2,
+    customerName: 'Lota Mongeskar',
+    rating: 3,
+    feedback: "4 The product not only met but exceeded my expectations. I'll definitely be a returning customer!"
+  },
+  {
+    imgURL: customer2,
+    customerName: 'Lota Mongeskar',
+    rating: 2.5,
+    feedback: "5 The product not only met but exceeded my expectations. I'll definitely be a returning customer!"
+  }
+];
+
+
+export const footerLinks = [
+  {
+    title: "Products",
+    links: [
+      { name: "Air Force 1", link: "/" },
+      { name: "Air Max 1", link: "/" },
+      { name: "Air Jordan 1", link: "/" },
+      { name: "Air Force 2", link: "/" },
+      { name: "Nike Waffle Racer", link: "/" },
+      { name: "Nike Cortez", link: "/" },
+    ],
+  },
+  {
+    title: "Help",
+    links: [
+      { name: "About us", link: "/" },
+      { name: "FAQs", link: "/" },
+      { name: "How it works", link: "/" },
+      { name: "Privacy policy", link: "/" },
+      { name: "Payment policy", link: "/" },
+    ],
+  },
+  {
+    title: "Get in touch",
+    links: [
+      { name: "customer@nike.com", link: "mailto:customer@nike.com" },
+      { name: "+92554862354", link: "tel:+92554862354" },
+    ],
+  },
+];
+
+export const socialMedia = [
+  { imgURL: <TwitterIcon className={"w-6 h-6 fill-black"}/>, 
+    label: "facebook logo" 
+  },
+  { imgURL: <FacebookIcon className={"w-6 h-6 fill-black"}/>, 
+    label: "twitter logo" 
+  },
+  { imgURL: <InstagramIcon className={"w-6 h-6 fill-black"}/>, 
+    label: "instagram logo" 
+  },
+];
