+++ src/sections/Footer.jsx
@@ -1,6 +1,6 @@
-import { copyrightSign } from "../assets/icons";
+import { CopyrightIcon } from "../assets/allIcons";
 import { footerLogo } from "../assets/images";
-import { footerLinks, socialMedia } from "../constants";
+import { footerLinks, socialMedia } from "../constants/data";
 
 const Footer = () => {
   return (
@@ -24,9 +24,9 @@ const Footer = () => {
             {socialMedia.map((icon) => (
               <div
                 className='flex justify-center items-center w-12 h-12 bg-white rounded-full'
-                key={icon.alt}
+                key={icon.label}
               >
-                <img src={icon.src} alt={icon.alt} width={24} height={24} />
+                {icon.imgURL}
               </div>
             ))}
           </div>
@@ -55,13 +55,7 @@ const Footer = () => {
 
       <div className='flex justify-between text-white-400 mt-24 max-sm:flex-col max-sm:items-center'>
         <div className='flex flex-1 justify-start items-center gap-2 font-montserrat cursor-pointer'>
-          <img
-            src={copyrightSign}
-            alt='copyright sign'
-            width={20}
-            height={20}
-            className='rounded-full m-0'
-          />
+          <CopyrightIcon className={"w-6 h-6 fill-transparent text-white"}/>
           <p>Copyright. All rights reserved.</p>
         </div>
         <p className='font-montserrat cursor-pointer'>Terms & Conditions</p>
