+++ src/sections/Services.jsx
@@ -1,10 +1,11 @@
-import { services } from "../constants";
 import { ServiceCard } from "../components";
+import { ourServices } from "../constants/data";
 
 const Services = () => {
+
   return (
     <section className='max-container flex justify-center flex-wrap gap-9'>
-      {services.map((service) => (
+      {ourServices.map((service) => (
         <ServiceCard key={service.label} {...service} />
       ))}
     </section>
